requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
prometheus-client==0.16.0

python-binance==1.0.18
ccxt==4.2.79
aiohttp==3.9.5

pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
tensorflow==2.18.1
ta==0.11.0

python-dotenv==1.0.0
requests==2.31.0
